+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Add UI to switch between play, out-jar, and package
 * Make mod description box scrollable for long text
 * Make Steam icon on mod open the mod's workshop page when clicked
+* Display multiple icons for mods that have local and workshop versions
 
 #### v3.30.3 ####
 * Fix package information being lost for patched classes (Alchyr)